fixed shit so it compiles but got a black screen fdsa

+++ include/maths/vec.h
@@ -133,8 +133,8 @@ public:
     }
 };
 
-typedef vec_t<float, 2> vec2_t;
-typedef vec_t<float, 3> vec3_t;
-typedef vec_t<float, 4> vec4_t;
+typedef vec_t<double, 2> vec2_t;
+typedef vec_t<double, 3> vec3_t;
+typedef vec_t<double, 4> vec4_t;
 
 #endif
